arm64: dts: ti: Add support for Kontron SMARC-sAM67
The Kontron SMARC-sAM67 is a SMARC module which features a TI AM67 SoC.
It supports the following features:
 * Quad-core AM67A94 with 1.4GHz
 * 8 GiB RAM
 * 64 GiB eMMC
 * 4 MiB SPI flash for failsafe booting
 * 4x UART
 * 1x USB 2.0
 * 1x USB 3.2 (or 4x USB3.2 with onboard USB 3.2 hub)
 * 1x RTC
 * 2x GBE
 * 1x QSPI (with 2 chip selects)
 * 1x SPI (with 2 chip selects)
 * 7x I2C
 * 4x CSI (*)
 * 2x LVDS (or one dual-link LVDS)
 * 1x DSI (*)
 * 1x DP (*)
 * onboard microcontroller for boot control, failsafe booting and
   external watchdog

(*) not yet supported by the kernel

There is a base device tree and overlays which will add optional
features. At the moment there is one full featured variant of that
board whose device tree is generated during build by merging all the
device tree overlays.

Signed-off-by: Michael Walle <mwalle@kernel.org>