drm/tidss: Remove max_pclk_khz from tidss display features
TIDSS hardware by itself does not have variable max_pclk for each VP.
The maximum pixel clock is determined by the limiting factor between
the functional clock and the PLL (parent to the VP/pixel clock).

The limitation that has been modeled till now comes from the clock
(PLL can only be programmed to a particular max value). Instead of
putting it as a constant field in dispc_features, we can query the
DM to see if requested clock can be set or not and use it in
"mode_valid()".

Replace constant "max_pclk_khz" in dispc_features with "curr_max_pclk"
in tidss_device structure which would be modified in runtime.
In mode_valid() call, check if a best frequency match for mode clock
can be found or not using "clk_round_rate()". Based on that, propagate
"cur_max_pclk" and query DM again only if the requested mode clock
is greater than cur_max_pclk. (As the preferred display mode is usually
the max resolution, driver ends up checking the highest clock the first
time itself which is used in subsequent checks)

Since TIDSS display controller provides clock tolerance of 5%, we use
this while checking the curr_max_pclk. Also, move up "dispc_pclk_diff()"
before it is called.

This will make the existing compatibles reusable if DSS features are
same across two SoCs with the only difference being the pixel clock.

Fixes: 7246e0929945 ("drm/tidss: Add OLDI bridge support")
Reviewed-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>